Renamed storyOpts to ItemConverter
refactored item and storyopts. storyopts renamed to itemConverter and convertToStory method moved from item struct to itemConverter struct and renamed to convert along with countComments and calculatePoints.
This means item is now much simpler and does not depend on anything else and doesn't have knowledge of other types of items like story. Also conversion is all contained in one struct now, the itemconverter, which we can easily and cleanly extend